default to disable writing min and max statistics for byte array
Updates the writer to default to not write min and max statistics for byte array. Adds an option to enable it.
Usually, byte arrays are representation of raw data that does not need statistics of min and max value. Storing such value in the parquet file will create unnecessary overhead and could result in large memory consumption.